Initial commit. Introduce InventoryShippingAdminUi Module
