--- app/code/Magento/InventoryShipping/Controller/Adminhtml/SourceSelection/Index.php
@@ -1,37 +1,0 @@
-<?php
-/**
- * Copyright © Magento, Inc. All rights reserved.
- * See COPYING.txt for license details.
- */
-declare(strict_types=1);
-
-namespace Magento\InventoryShipping\Controller\Adminhtml\SourceSelection;
-
-use Magento\Backend\App\Action;
-use Magento\Backend\App\Action\Context;
-use Magento\Framework\Registry;
-use Magento\Sales\Model\OrderRepository;
-use Magento\Framework\Exception\InputException;
-use Magento\Framework\Exception\NoSuchEntityException;
-
-/**
- * Class Index
- */
-class Index extends Action
-{
-    /**
-     * @see _isAllowed()
-     */
-    const ADMIN_RESOURCE = 'Magento_InventoryApi::source';
-
-    /**
-     * @inheritdoc
-     */
-    public function execute()
-    {
-        $this->_view->loadLayout();
-        $this->_setActiveMenu('Magento_Sales::sales_order');
-        $this->_view->getPage()->getConfig()->getTitle()->prepend(__('Source Selection'));
-        $this->_view->renderLayout();
-    }
-}
