--- app/code/Magento/InventoryShipping/Ui/DataProvider/GetSourcesByStockIdSkuAndQty.php
@@ -1,121 +1,0 @@
-<?php
-/**
- * Copyright © Magento, Inc. All rights reserved.
- * See COPYING.txt for license details.
- */
-declare(strict_types=1);
-
-namespace Magento\InventoryShipping\Ui\DataProvider;
-
-use Magento\InventorySourceSelectionApi\Api\Data\ItemRequestInterfaceFactory;
-use Magento\InventorySourceSelectionApi\Api\Data\InventoryRequestInterfaceFactory;
-use Magento\InventorySourceSelectionApi\Api\SourceSelectionServiceInterface;
-use Magento\InventorySourceSelectionApi\Api\GetDefaultSourceSelectionAlgorithmCodeInterface;
-use Magento\InventoryApi\Api\SourceRepositoryInterface;
-
-class GetSourcesByStockIdSkuAndQty
-{
-    /**
-     * @var ItemRequestInterfaceFactory
-     */
-    private $itemRequestFactory;
-
-    /**
-     * @var InventoryRequestInterfaceFactory
-     */
-    private $inventoryRequestFactory;
-
-    /**
-     * @var SourceSelectionServiceInterface
-     */
-    private $sourceSelectionService;
-
-    /**
-     * @var GetDefaultSourceSelectionAlgorithmCodeInterface
-     */
-    private $getDefaultSourceSelectionAlgorithmCode;
-
-    /**
-     * @var SourceRepositoryInterface
-     */
-    private $sourceRepository;
-
-    /**
-     * @var array
-     */
-    private $sources = [];
-
-    /**
-     * @param ItemRequestInterfaceFactory $itemRequestFactory
-     * @param InventoryRequestInterfaceFactory $inventoryRequestFactory
-     * @param SourceSelectionServiceInterface $sourceSelectionService
-     * @param GetDefaultSourceSelectionAlgorithmCodeInterface $getDefaultSourceSelectionAlgorithmCode
-     * @param SourceRepositoryInterface $sourceRepository
-     */
-    public function __construct(
-        ItemRequestInterfaceFactory $itemRequestFactory,
-        InventoryRequestInterfaceFactory $inventoryRequestFactory,
-        SourceSelectionServiceInterface $sourceSelectionService,
-        GetDefaultSourceSelectionAlgorithmCodeInterface $getDefaultSourceSelectionAlgorithmCode,
-        SourceRepositoryInterface $sourceRepository
-    ) {
-        $this->itemRequestFactory = $itemRequestFactory;
-        $this->inventoryRequestFactory = $inventoryRequestFactory;
-        $this->sourceSelectionService = $sourceSelectionService;
-        $this->getDefaultSourceSelectionAlgorithmCode = $getDefaultSourceSelectionAlgorithmCode;
-        $this->sourceRepository = $sourceRepository;
-    }
-
-    /**
-     * @param int $stockId
-     * @param string $sku
-     * @param float $qty
-     * @return array
-     * @throws \Magento\Framework\Exception\NoSuchEntityException
-     */
-    public function execute(int $stockId, string $sku, float $qty): array
-    {
-        $algorithmCode = $this->getDefaultSourceSelectionAlgorithmCode->execute();
-
-        $requestItem = $this->itemRequestFactory->create([
-            'sku' => $sku,
-            'qty' => $qty
-        ]);
-        $inventoryRequest = $this->inventoryRequestFactory->create([
-            'stockId' => $stockId,
-            'items' => [$requestItem]
-        ]);
-        $sourceSelectionResult = $this->sourceSelectionService->execute(
-            $inventoryRequest,
-            $algorithmCode
-        );
-        $result = [];
-        foreach ($sourceSelectionResult->getSourceSelectionItems() as $item) {
-            $sourceCode = $item->getSourceCode();
-            $result[] = [
-                'sourceName' => $this->getSourceName($sourceCode),
-                'sourceCode' => $sourceCode,
-                'qtyAvailable' => $item->getQtyAvailable(),
-                'qtyToDeduct' => $item->getQtyToDeduct()
-            ];
-        }
-
-        return $result;
-    }
-
-    /**
-     * Get source name by code
-     *
-     * @param string $sourceCode
-     * @return mixed
-     * @throws \Magento\Framework\Exception\NoSuchEntityException
-     */
-    private function getSourceName(string $sourceCode): string
-    {
-        if (!isset($this->sources[$sourceCode])) {
-            $this->sources[$sourceCode] = $this->sourceRepository->get($sourceCode)->getName();
-        }
-
-        return $this->sources[$sourceCode];
-    }
-}
