+++ app/code/Magento/InventoryShipping/composer.json
@@ -4,16 +4,13 @@
     "require": {
         "php": "~7.1.3||~7.2.0",
         "magento/framework": "*",
-        "magento/module-backend": "*",
         "magento/module-inventory-api": "*",
         "magento/module-inventory-catalog-api": "*",
         "magento/module-inventory-configuration-api": "*",
         "magento/module-inventory-sales-api": "*",
         "magento/module-inventory-source-selection-api": "*",
         "magento/module-sales": "*",
-        "magento/module-shipping": "*",
-        "magento/module-store": "*",
-        "magento/module-ui": "*"
+        "magento/module-store": "*"
     },
     "type": "magento2-module",
     "license": [
