--- app/code/Magento/InventoryShipping/etc/adminhtml/di.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0"?>
-<!--
-/**
- * Copyright © Magento, Inc. All rights reserved.
- * See COPYING.txt for license details.
- */
--->
-<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:ObjectManager/etc/config.xsd">
-    <!-- Buttons configurations -->
-    <type name="Magento\InventoryShipping\Ui\Component\Control\SourceSelection\AlgorithmSelectionButton">
-        <arguments>
-            <argument name="targetName" xsi:type="string">inventory_shipping_source_selection_form.inventory_shipping_source_selection_form</argument>
-        </arguments>
-    </type>
-    <type name="Magento\Sales\Block\Adminhtml\Order\Invoice\Create\Form">
-        <plugin name="disallow_create_shipment_in_multi_source_mode" type="Magento\InventoryShipping\Plugin\Sales\Block\Order\Invoice\Create\DisallowCreateShipmentPlugin"/>
-    </type>
-
-    <type name="Magento\Shipping\Controller\Adminhtml\Order\Shipment\NewAction">
-        <arguments>
-            <argument name="shipmentProvider" xsi:type="object">Magento\InventoryShipping\Model\ShipmentProvider</argument>
-        </arguments>
-    </type>
-</config>
