--- app/code/Magento/InventoryShipping/etc/adminhtml/routes.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0"?>
-<!--
-/**
- * Copyright © Magento, Inc. All rights reserved.
- * See COPYING.txt for license details.
- */
--->
-<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:App/etc/routes.xsd">
-    <router id="admin">
-        <route id="inventoryshipping" frontName="inventoryshipping">
-            <module name="Magento_InventoryShipping" before="Magento_Backend" />
-        </route>
-    </router>
-</config>
