+++ app/code/Magento/InventoryShipping/etc/events.xml
@@ -12,7 +12,4 @@
     <event name="sales_order_invoice_save_after">
         <observer name="inventory_sales_virtual_source_deduction_processor" instance="Magento\InventoryShipping\Observer\VirtualSourceDeductionProcessor"/>
     </event>
-    <event name="controller_action_predispatch_adminhtml_order_shipment_new">
-        <observer name="inventory_shipping_check_data_before_load_new_shipment" instance="Magento\InventoryShipping\Observer\NewShipmentLoadBefore"/>
-    </event>
 </config>
