--- app/code/Magento/InventoryShipping/view/adminhtml/layout/adminhtml_order_shipment_new.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0"?>
-<page xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:View/Layout/etc/page_configuration.xsd">
-    <body>
-        <referenceBlock name="form">
-            <block class="Magento\InventoryShipping\Block\Adminhtml\Shipment\Inventory" name="inventory_shipment"
-                   template="Magento_InventoryShipping::shipment/inventory.phtml" before="-" />
-        </referenceBlock>
-    </body>
-</page>
