--- app/code/Magento/InventoryShipping/view/adminhtml/layout/sales_order_view.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0"?>
-<!--
-/**
- * Copyright © Magento, Inc. All rights reserved.
- * See COPYING.txt for license details.
- */
--->
-<page xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" layout="admin-2columns-left" xsi:noNamespaceSchemaLocation="urn:magento:framework:View/Layout/etc/page_configuration.xsd">
-    <body>
-        <referenceContainer name="content">
-            <block class="Magento\InventoryShipping\Block\Adminhtml\Order\View\ShipButton" name="sales_order_view_ship_button"/>
-        </referenceContainer>
-    </body>
-</page>
