+++ app/code/Magento/InventoryShippingAdminUi/README.md
@@ -1,0 +1,1 @@
+# InventoryShippingAdminUi
