+++ app/code/Magento/InventoryShippingAdminUi/Ui/DataProvider/SourceSelectionDataProvider.php
@@ -1,0 +1,229 @@
+<?php
+/**
+ * Copyright © Magento, Inc. All rights reserved.
+ * See COPYING.txt for license details.
+ */
+declare(strict_types=1);
+
+namespace Magento\InventoryShippingAdminUi\Ui\DataProvider;
+
+use Magento\Framework\Api\Filter;
+use Magento\Framework\Exception\LocalizedException;
+use Magento\Framework\Exception\NoSuchEntityException;
+use Magento\Sales\Model\OrderRepository;
+use Magento\Sales\Model\Order\Item;
+use Magento\Ui\DataProvider\AbstractDataProvider;
+use Magento\InventorySalesApi\Model\StockByWebsiteIdResolverInterface;
+use Magento\InventoryConfigurationApi\Api\GetStockItemConfigurationInterface;
+use Magento\Framework\App\RequestInterface;
+use Magento\InventoryShippingAdminUi\Ui\DataProvider\GetSourcesByStockIdSkuAndQty;
+
+class SourceSelectionDataProvider extends AbstractDataProvider
+{
+    /**
+     * @var RequestInterface
+     */
+    private $request;
+
+    /**
+     * @var OrderRepository
+     */
+    private $orderRepository;
+
+    /**
+     * @var StockByWebsiteIdResolverInterface
+     */
+    private $stockByWebsiteIdResolver;
+
+    /**
+     * @var GetStockItemConfigurationInterface
+     */
+    private $getStockItemConfiguration;
+
+    /**
+     * @var GetSourcesByStockIdSkuAndQty
+     */
+    private $getSourcesByStockIdSkuAndQty;
+
+    /**
+     * @var array
+     */
+    private $sources = [];
+
+    /**
+     * @param string $name
+     * @param string $primaryFieldName
+     * @param string $requestFieldName
+     * @param RequestInterface $request
+     * @param OrderRepository $orderRepository
+     * @param StockByWebsiteIdResolverInterface $stockByWebsiteIdResolver
+     * @param GetStockItemConfigurationInterface $getStockItemConfiguration
+     * @param GetSourcesByStockIdSkuAndQty $getSourcesByStockIdSkuAndQty
+     * @param array $meta
+     * @param array $data
+     * @SuppressWarnings(PHPMD.ExcessiveParameterList)
+     */
+    public function __construct(
+        $name,
+        $primaryFieldName,
+        $requestFieldName,
+        RequestInterface $request,
+        OrderRepository $orderRepository,
+        StockByWebsiteIdResolverInterface $stockByWebsiteIdResolver,
+        GetStockItemConfigurationInterface $getStockItemConfiguration,
+        GetSourcesByStockIdSkuAndQty $getSourcesByStockIdSkuAndQty,
+        array $meta = [],
+        array $data = []
+    ) {
+        parent::__construct($name, $primaryFieldName, $requestFieldName, $meta, $data);
+        $this->request = $request;
+        $this->orderRepository = $orderRepository;
+        $this->stockByWebsiteIdResolver = $stockByWebsiteIdResolver;
+        $this->getStockItemConfiguration = $getStockItemConfiguration;
+        $this->getSourcesByStockIdSkuAndQty = $getSourcesByStockIdSkuAndQty;
+    }
+
+    /**
+     * Disable for collection processing | ????
+     *
+     * @param Filter $filter
+     * @return bool
+     * @SuppressWarnings(PHPMD.UnusedFormalParameter)
+     */
+    public function addFilter(Filter $filter)
+    {
+        return null;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function getData()
+    {
+        $data = [];
+        $orderId = $this->request->getParam('order_id');
+        /** @var \Magento\Sales\Model\Order $order */
+        $order = $this->orderRepository->get($orderId);
+        $websiteId = $order->getStore()->getWebsiteId();
+        $stockId = (int)$this->stockByWebsiteIdResolver->execute((int)$websiteId)->getStockId();
+
+        foreach ($order->getAllItems() as $orderItem) {
+            if ($orderItem->getIsVirtual()
+                || $orderItem->getLockedDoShip()
+                || $orderItem->getHasChildren()) {
+                continue;
+            }
+
+            $orderItemId = $orderItem->getId();
+            //TODO: Need to add additional logic for bundle product with flag ship Together
+            if ($orderItem->getParentItem() && !$orderItem->isShipSeparately()) {
+                $orderItemId = $orderItem->getParentItemId();
+            }
+
+            $qty = $orderItem->getSimpleQtyToShip();
+            $qty = $this->castQty($orderItem, $qty);
+            $sku = $orderItem->getSku();
+            $data[$orderId]['items'][] = [
+                'orderItemId' => $orderItemId,
+                'sku' => $sku,
+                'product' => $this->getProductName($orderItem),
+                'qtyToShip' => $qty,
+                'sources' => $this->getSources($stockId, $sku, $qty),
+                'isManageStock' => $this->isManageStock($sku, $stockId)
+            ];
+        }
+        $data[$orderId]['websiteId'] = $websiteId;
+        $data[$orderId]['order_id'] = $orderId;
+        foreach ($this->sources as $code => $name) {
+            $data[$orderId]['sourceCodes'][] = [
+                'value' => $code,
+                'label' => $name
+            ];
+        }
+
+        return $data;
+    }
+
+    /**
+     * @param int $stockId
+     * @param string $sku
+     * @param float $qty
+     * @return array
+     * @throws NoSuchEntityException
+     */
+    private function getSources(int $stockId, string $sku, float $qty): array
+    {
+        $sources = $this->getSourcesByStockIdSkuAndQty->execute($stockId, $sku, $qty);
+        foreach ($sources as $source) {
+            $this->sources[$source['sourceCode']] = $source['sourceName'];
+        }
+        return $sources;
+    }
+
+    /**
+     * @param $itemSku
+     * @param $stockId
+     * @return bool
+     * @throws LocalizedException
+     */
+    private function isManageStock($itemSku, $stockId)
+    {
+        $stockItemConfiguration = $this->getStockItemConfiguration->execute($itemSku, $stockId);
+        if (!empty($stockItemConfiguration)) {
+            return $stockItemConfiguration->isManageStock();
+        }
+
+        return true;
+    }
+
+    /**
+     * Generate display product name
+     * @param Item $item
+     * @return null|string
+     */
+    private function getProductName(Item $item)
+    {
+        //TODO: need to transfer this to html block and render on Ui
+        $name = $item->getName();
+        if ($parentItem = $item->getParentItem()) {
+            $name = $parentItem->getName();
+            $options = [];
+            if ($productOptions = $parentItem->getProductOptions()) {
+                if (isset($productOptions['options'])) {
+                    $options = array_merge($options, $productOptions['options']);
+                }
+                if (isset($productOptions['additional_options'])) {
+                    $options = array_merge($options, $productOptions['additional_options']);
+                }
+                if (isset($productOptions['attributes_info'])) {
+                    $options = array_merge($options, $productOptions['attributes_info']);
+                }
+                if (count($options)) {
+                    foreach ($options as $option) {
+                        $name .= '<dd>' . $option['label'] . ': ' . $option['value'] .'</dd>';
+                    }
+                } else {
+                    $name .= '<dd>' . $item->getName() . '</dd>';
+                }
+            }
+        }
+
+        return $name;
+    }
+
+    /**
+     * @param Item $item
+     * @param string|int|float $qty
+     * @return float|int
+     */
+    private function castQty(Item $item, $qty)
+    {
+        if ($item->getIsQtyDecimal()) {
+            $qty = (double)$qty;
+        } else {
+            $qty = (int)$qty;
+        }
+
+        return $qty > 0 ? $qty : 0;
+    }
+}
