+++ app/code/Magento/InventoryShippingAdminUi/composer.json
@@ -1,0 +1,30 @@
+{
+  "name": "magento/module-inventory-shipping-admin-ui",
+  "description": "N/A",
+  "require": {
+    "php": "~7.1.3||~7.2.0",
+    "magento/framework": "*",
+    "magento/module-backend": "*",
+    "magento/module-inventory-api": "*",
+    "magento/module-inventory-configuration-api": "*",
+    "magento/module-inventory-sales-api": "*",
+    "magento/module-inventory-source-selection-api": "*",
+    "magento/module-inventory-shipping": "*",
+    "magento/module-sales": "*",
+    "magento/module-shipping": "*",
+    "magento/module-ui": "*"
+  },
+  "type": "magento2-module",
+  "license": [
+    "OSL-3.0",
+    "AFL-3.0"
+  ],
+  "autoload": {
+    "files": [
+      "registration.php"
+    ],
+    "psr-4": {
+      "Magento\\InventoryShippingAdminUi\\": ""
+    }
+  }
+}
