+++ app/code/Magento/InventoryShippingAdminUi/etc/di.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0"?>
+<!--
+/**
+ * Copyright © Magento, Inc. All rights reserved.
+ * See COPYING.txt for license details.
+ */
+-->
+<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:ObjectManager/etc/config.xsd">
+    <type name="Magento\Shipping\Block\Adminhtml\Create">
+        <plugin name="back_button_url" type="Magento\InventoryShippingAdminUi\Plugin\Sales\Shipment\Block\Adminhtml\BackButtonUrlOnNewShipmentPagePlugin"/>
+    </type>
+</config>
