+++ app/code/Magento/InventoryShippingAdminUi/view/adminhtml/layout/inventoryshipping_sourceselection_index.xml
