+++ app/code/Magento/InventoryShippingAdminUi/view/adminhtml/layout/sales_order_view.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0"?>
+<!--
+/**
+ * Copyright © Magento, Inc. All rights reserved.
+ * See COPYING.txt for license details.
+ */
+-->
+<page xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" layout="admin-2columns-left" xsi:noNamespaceSchemaLocation="urn:magento:framework:View/Layout/etc/page_configuration.xsd">
+    <body>
+        <referenceContainer name="content">
+            <block class="Magento\InventoryShippingAdminUi\Block\Adminhtml\Order\View\ShipButton" name="sales_order_view_ship_button"/>
+        </referenceContainer>
+    </body>
+</page>
