+++ composer.json
@@ -192,6 +192,7 @@
         "magento/module-inventory-sales-admin-ui": "*",
         "magento/module-inventory-sales-api": "*",
         "magento/module-inventory-shipping": "*",
+        "magento/module-inventory-shipping-admin-ui": "*",
         "magento/module-inventory-source-selection-api": "*",
         "magento/module-inventory-source-selection": "*",
         "magento/module-layered-navigation": "*",
